BaseModel.Fit: Add Checkpoint Create Bool
When training, the first patience array figure is 1500 while the min_steps_to_checkpoint is 2000. This sets up a scenario where the the method tries to reset to a snapshot before a snapshot takes place. This commit tracks if checkpoint has been made and only allows a reset to happen if that's true.